WIP: Add Suggestions In WB Form
