+++ scout/server/blueprints/cases/views.py
@@ -32,6 +32,7 @@ from scout.server.utils import (
     user_institutes,
     zip_dir_to_obj,
 )
+from scout.utils.gene import parse_raw_gene_ids, parse_raw_gene_symbols
 
 from . import controllers
 
@@ -361,58 +362,6 @@ def phenotypes(institute_id, case_name, phenotype_id=None):
     return redirect("#".join([case_url, "phenotypes_panel"]))
 
 
-def parse_raw_gene_ids(raw_symbols):
-    """Parse raw gene symbols for hgnc_symbols from web form autocompletion.
-
-    Arguments:
-        raw_symbol_strings(list(string)) - formated "17284 | POT1 (hPot1, POT1)"
-
-    Returns:
-        hgnc_ids(set(int))
-    """
-    hgnc_ids = set()
-
-    for raw_symbol in raw_symbols:
-        LOG.debug("raw gene: {}".format(raw_symbol))
-        # avoid empty lists
-        if raw_symbol:
-            # take the first nubmer before |, and remove any space.
-            try:
-                hgnc_ids.add(int(raw_symbol.split("|", 1)[0].replace(" ", "")))
-            except ValueError:
-                flash(
-                    "Provided gene info could not be parsed! "
-                    "Please allow autocompletion to finish.",
-                    "warning",
-                )
-
-    LOG.debug("Parsed HGNC symbols {}".format(hgnc_ids))
-
-    return hgnc_ids
-
-
-def parse_raw_gene_symbols(raw_symbols_list):
-    """Parse list of concatenated gene symbol list for hgnc_symbols from Phenomizer.
-
-    Arguments:
-        raw_symbols(list(string)) - e.g. ("POT1 | MUTYH", "POT1 | ATXN1 | ATXN7")
-
-    Returns:
-        hgnc_symbols(set(string)) - set of (unique) gene symbols without intervening chars
-    """
-    hgnc_symbols = set()
-
-    for raw_symbols in raw_symbols_list:
-        LOG.debug("raw gene: {}".format(raw_symbols))
-        # avoid empty lists
-        if raw_symbols:
-            hgnc_symbols.update(
-                raw_symbol.split(" ", 1)[0] for raw_symbol in raw_symbols.split("|")
-            )
-    LOG.debug("Parsed HGNC symbols {}".format(hgnc_symbols))
-
-    return hgnc_symbols
-
 
 @cases_bp.route("/<institute_id>/<case_name>/phenotypes/actions", methods=["POST"])
 def phenotypes_actions(institute_id, case_name):
@@ -444,7 +393,15 @@ def phenotypes_actions(institute_id, case_name):
             store.remove_phenotype(institute_obj, case_obj, user_obj, case_url, hpo_id)
 
     if action == "ADDGENE":
-        hgnc_ids = parse_raw_gene_ids(request.form.getlist("genes"))
+        try:
+            hgnc_ids = parse_raw_gene_ids(request.form.getlist("genes"))
+        except ValueError:
+                        flash(
+                    "Provided gene info could not be parsed! "
+                    "Please allow autocompletion to finish.",
+                    "warning",
+                )
+        #TODO: dont crash
         store.update_dynamic_gene_list(case_obj, hgnc_ids=list(hgnc_ids), add_only=True)
 
     if action == "GENES":
