+++ scout/server/blueprints/panels/templates/panels/panels.html
@@ -39,11 +39,11 @@
       <form action="{{ url_for('panels.panels') }}" enctype="multipart/form-data" method="POST">
         <div class="row">
           <div class="col-5">
-            <input type="text" name="search_for" class="form-control typeahead_gene" data-provide="typeahead" autocomplete="off" placeholder="Search Gene in Panels" required
-                   pattern="^[a-zA-Z0-9_\-]*$" title="Only alphanumeric characters (A-Z+a-z+0-9) and underscores allowed.">
+            <input type="text" name="search_for" class="form-control typeahead_gene mb-1" data-provide="typeahead" autocomplete="off" placeholder="Search Gene in Panels" required
+                   pattern="^[0-9]+\s*\|\s*.*" title="Only alphanumeric characters (A-Z+a-z+0-9) and underscores allowed.">
           </div>
           <div class="col">
-            <button type="submit" class="btn btn-secondary">Search!</button>
+            <button type="submit" name="action" value="ADDGENE" class="btn btn-secondary">Search!</button>
           </div>
           <div class="col">
           </div>
